some more compile fixes

+++ src/Win95/std.c
@@ -32,7 +32,7 @@ size_t stdFileRead(FILE* _File, void* _DstBuf, size_t _Count)
 }
 
 // 0x0048c660
-char* stdFileGets(FILE* _File, char* _Buf, int _MaxCount)
+char* stdFileGets(FILE* _File, char* _Buf, size_t _MaxCount)
 {
     return fgets(_Buf, _MaxCount, _File);
 }
@@ -44,13 +44,13 @@ size_t stdFileWrite(FILE* _File, const void* _Str, size_t _Count)
 }
 
 // 0x0048c6b0
-long stdFtell(FILE* _File)
+int stdFtell(FILE* _File)
 {
     return ftell(_File);
 }
 
 // 0x0048c6c0
-int stdFseek(FILE* _File, long _Offset, int _Origin)
+int stdFseek(FILE* _File, int _Offset, int _Origin)
 {
     return fseek(_File, _Offset, _Origin);
 }
@@ -79,7 +79,7 @@ int stdFilePrintf(FILE* f, const char* format, ...)
 }
 
 // 0x0048c680
-wchar_t* stdFileGetws(FILE* _File, wchar_t* _Dst, int _SizeInWords)
+wchar_t* stdFileGetws(FILE* _File, wchar_t* _Dst, size_t _SizeInWords)
 {
     return fgetws(_Dst, _SizeInWords, _File);
 }
